feat(Event): swap to TS
